+++ contracts/IVoucherHub.sol
@@ -15,7 +15,7 @@ interface IVoucherHub {
         uint256 voucherType,
         uint256 value
     );
-    event VoucherDebited(address indexed voucher, uint256 debitedAmount);
+    event VoucherDebited(address indexed voucher, uint256 sponsoredAmount);
     event VoucherTypeCreated(uint256 indexed id, string description, uint256 duration);
     event VoucherTypeDescriptionUpdated(uint256 indexed id, string description);
     event VoucherTypeDurationUpdated(uint256 indexed id, uint256 duration);
@@ -28,7 +28,15 @@ interface IVoucherHub {
         uint256 value
     ) external returns (address voucherAddress);
 
-    function debitVoucher(uint256 debitAmount) external;
+    function debitVoucher(
+        uint256 voucherTypeId,
+        address app,
+        uint256 appPrice,
+        address dataset,
+        uint256 datasetPrice,
+        address workerpool,
+        uint256 workerpoolPrice
+    ) external returns (uint256 sponsoredAmount);
 
     function getVoucher(address account) external view returns (address);