+++ contracts/beacon/Voucher.sol
@@ -3,6 +3,8 @@
 
 pragma solidity ^0.8.20;
 
+import {IexecLibOrders_v5} from "@iexec/poco/contracts/libs/IexecLibOrders_v5.sol";
+import {IexecPoco1} from "@iexec/poco/contracts/modules/interfaces/IexecPoco1.v8.sol";
 import {OwnableUpgradeable} from "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {IVoucherHub} from "../IVoucherHub.sol";
@@ -19,6 +21,7 @@ contract Voucher is OwnableUpgradeable, IVoucher {
         uint256 _expiration;
         uint256 _type;
         mapping(address => bool) _authorizedAccounts;
+        mapping(bytes32 dealId => uint256) _sponsoredAmounts;
     }
 
     // keccak256(abi.encode(uint256(keccak256("iexec.voucher.storage.Voucher")) - 1))
@@ -59,6 +62,58 @@ contract Voucher is OwnableUpgradeable, IVoucher {
     }
 
     /**
+     * Match orders on Poco. Eligible assets prices will be debited from the
+     * voucher if possible, then non-sponsored amount will be debited from the
+     * iExec account of the requester.
+     *
+     * @param appOrder The app order.
+     * @param datasetOrder The dataset order.
+     * @param workerpoolOrder The workerpool order.
+     * @param requestOrder The request order.
+     */
+    function matchOrders(
+        IexecLibOrders_v5.AppOrder calldata appOrder,
+        IexecLibOrders_v5.DatasetOrder calldata datasetOrder,
+        IexecLibOrders_v5.WorkerpoolOrder calldata workerpoolOrder,
+        IexecLibOrders_v5.RequestOrder calldata requestOrder
+    ) external returns (bytes32 dealId) {
+        uint256 appPrice = appOrder.appprice;
+        uint256 datasetPrice = datasetOrder.datasetprice;
+        uint256 workerpoolPrice = workerpoolOrder.workerpoolprice;
+        VoucherStorage storage $ = _getVoucherStorage();
+        IVoucherHub voucherHub = IVoucherHub($._voucherHub);
+        uint256 sponsoredAmount = voucherHub.debitVoucher(
+            $._type,
+            appOrder.app,
+            appPrice,
+            datasetOrder.dataset,
+            datasetPrice,
+            workerpoolOrder.workerpool,
+            workerpoolPrice
+        );
+        uint256 dealPrice = appPrice + datasetPrice + workerpoolPrice;
+        address iexecPoco = voucherHub.getIexecPoco();
+        if (sponsoredAmount != dealPrice) {
+            // Transfer non-sponsored amount from the iExec account of the
+            // requester to the iExec account of the voucher
+            IERC20(iexecPoco).transferFrom(
+                requestOrder.requester,
+                address(this),
+                dealPrice - sponsoredAmount
+            );
+        }
+        dealId = IexecPoco1(iexecPoco).sponsorMatchOrders(
+            appOrder,
+            datasetOrder,
+            workerpoolOrder,
+            requestOrder
+        );
+        $._sponsoredAmounts[dealId] = sponsoredAmount;
+        emit OrdersMatchedWithVoucher(dealId);
+        return dealId;
+    }
+
+    /**
      * Retrieve the address of the voucher hub associated with the voucher.
      * @return voucherHubAddress The address of the voucher hub.
      */
@@ -86,12 +141,22 @@ contract Voucher is OwnableUpgradeable, IVoucher {
     }
 
     /**
+     * Get amount sponsored in a deal.
+     * @param dealId The ID of the deal.
+     */
+    function getSponsoredAmount(bytes32 dealId) external view returns (uint256) {
+        VoucherStorage storage $ = _getVoucherStorage();
+        return $._sponsoredAmounts[dealId];
+    }
+
+    /**
      * Get voucher balance.
      */
     function getBalance() external view returns (uint256) {
         return IERC20(getVoucherHub()).balanceOf(address(this));
     }
 
+    // TODO: Move "authorize" functions for order consistency
     /**
      * Sets authorization for an account.
      * @param account The account to authorize.