+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## vNEXT
+- Match orders through voucher. (#12)
 - Add external-hardhat network configuration. (#11)
 - Add voucher credit and SRLC manipulation. (#10)
     - SRLC and iExec poco is mocked.